refactor: Extract `getChatIcon` into a shared utility and update chat-related components to use it.
